Carefully copy and delete objects
If there's a problem copying or deleting an object, log the error and
continue. If the error is one that can be retried, leave the message
on the queue; if the error won't be resolvable, indicate the problem
and continue.

In particular, if the source object to copy is no longer available,
there's nothing to do.